Finished PostService

+++ CS321_W5D2_BlogAPI.Core/Services/PostService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Security.Cryptography;
 using CS321_W5D2_BlogAPI.Core.Models;
 
 namespace CS321_W5D2_BlogAPI.Core.Services
@@ -19,11 +20,15 @@ namespace CS321_W5D2_BlogAPI.Core.Services
 
         public Post Add(Post newPost)
         {
-            // TODO: Prevent users from adding to a blog that isn't theirs
-            //     Use the _userService to get the current users id.
-            //     You may have to retrieve the blog in order to check user id
-            // TODO: assign the current date to DatePublished
-            return _postRepository.Add(newPost);
+            var currentUser = _userService.CurrentUserId;
+            var currentBlog = _blogRepository.Get(newPost.BlogId);
+            if(currentUser == currentBlog.UserId)
+            {
+                newPost.DatePublished = DateTime.Now;
+                return _postRepository.Add(newPost);
+            }
+
+            throw new Exception("Nuh uh uh!  Nuh uh uh! Not yours!"); 
         }
 
         public Post Get(int id)
@@ -44,14 +49,22 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         public void Remove(int id)
         {
             var post = this.Get(id);
-            // TODO: prevent user from deleting from a blog that isn't theirs
-            _postRepository.Remove(id);
+            if (_userService.CurrentUserId == post.Blog.UserId)
+            {
+                _postRepository.Remove(id);
+                return;
+            }
+            throw new Exception("Nuh uh uh!  Nuh uh uh! Not yours!");
         }
 
         public Post Update(Post updatedPost)
         {
-            // TODO: prevent user from updating a blog that isn't theirs
-            return _postRepository.Update(updatedPost);
+            var currentUser = _userService.CurrentUserId;
+            if (currentUser == updatedPost.Blog.UserId)
+            {
+                return _postRepository.Update(updatedPost);
+            }
+            throw new Exception("Nuh uh uh!  Nuh uh uh! Not yours!");
         }
 
     }
